# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/app/approved-exam/head-teacher/{id}": {
      "get": {
        "operationId": "ApprovedExamController_listApprovedByHeadTeacher",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "ApprovedExam"
        ]
      }
    },
...
    "/app/teacher": {
      "post": {
        "operationId": "TeacherController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeacherDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Teacher"
        ]
      },
      "get": {
        "operationId": "TeacherController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Teacher"
        ]
      }
    },
...
    "/app/teacher/all/{id}": {
      "get": {
        "operationId": "TeacherController_findOneAll",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Teacher"
        ]
      }
    },
    "/app/teacher/deleted/{id}": {
      "get": {
        "operationId": "TeacherController_findOneDeleted",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Teacher"
        ]
      }
    },
    "/app/teacher/{id}": {
      "patch": {
        "operationId": "TeacherController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTeacherDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Teacher"
        ]
      },
      "delete": {
        "operationId": "TeacherController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Teacher"
        ]
      },
      "get": {
        "operationId": "TeacherController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Teacher"
        ]
      }
    },
    "/app/teacher/restore/{id}": {
      "patch": {
        "operationId": "TeacherController_restore",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Teacher"
        ]
      }
    },
    "/app/teacher/{id}/subjects": {
      "get": {
        "operationId": "TeacherController_getSubjects",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Teacher"
        ]
      }
    },
...
    "/app/head-teacher": {
      "post": {
        "operationId": "HeadTeacherController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHeadTeacherDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "HeadTeacher"
        ]
      },
      "get": {
        "operationId": "HeadTeacherController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "HeadTeacher"
        ]
      }
    },
...
    "/app/head-teacher/all/{id}": {
      "get": {
        "operationId": "HeadTeacherController_findOneAll",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "HeadTeacher"
        ]
      }
    },
    "/app/head-teacher/deleted/{id}": {
      "get": {
        "operationId": "HeadTeacherController_findOneDeleted",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "HeadTeacher"
        ]
      }
    },
    "/app/head-teacher/{id}": {
      "get": {
        "operationId": "HeadTeacherController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "HeadTeacher"
        ]
      },
      "patch": {
        "operationId": "HeadTeacherController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHeadTeacherDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "HeadTeacher"
        ]
      },
      "delete": {
        "operationId": "HeadTeacherController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "HeadTeacher"
        ]
      }
    },
    "/app/head-teacher/restore/{id}": {
      "patch": {
        "operationId": "HeadTeacherController_restore",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "HeadTeacher"
        ]
      }
    },
...
  },
  "info": {
    "title": "Script",
    "description": "API description",
    "version": "0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateTeacherDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateTeacherDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateHeadTeacherDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateHeadTeacherDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}